refactor(model): split EventStore occurrence-record API into extension
Extract the 360-line "Calendar Feedback / Logs (Occurrence-scoped)"
plus "Generic Record Pruning" sections out of EventStore.swift into
EventStore+OccurrenceRecords.swift -- pure file reorganization,
no logic change. The new extension covers:

* feedbackRecord / upsertFeedbackRecord
* logRecord / upsertLogRecord
* syncCalendarEventColorDepthIfNeeded (now fileprivate to the extension)
* appendTimelineNote / updateTimelineNote / deleteTimelineNote
* deleteLogRecord / prefilledDraft / setChatConversationID
* pruneRecords<T> generic helpers (fileprivate)
* prune{Feedback,LogRecords}For{Deleted,Recurring}* public wrappers
* calendarOccurrenceKey(for:) helper (fileprivate)

Access changes in main EventStore.swift:
* The @Published state properties drop `private(set)` so the
  extension in another file can mutate them. External readers
  (SwiftUI views, services) still read via the same paths; the
  rule is "go through the dedicated mutation helpers" -- written
  as a comment above the declarations.
* saveCalendarEventFeedbackRecords() and saveCalendarEventLogRecords()
  drop `private` so the extension can call them.
* saveCalendarEvents(refreshInterrupts:) drops `private` for the
  same reason (called from syncCalendarEventColorDepthIfNeeded which
  now lives in the extension).

Result: main file 1453 -> 1111 lines (-342). Extension file 360
lines. Same MainActor isolation (inherited by extension). Build
clean. New file's path is quoted in pbxproj because the `+` glyph
is not legal in unquoted OpenStep strings.
